+++ packages/server/controllers/apps.js
@@ -5,7 +5,7 @@ import Ajv from 'ajv';
 import crypto from 'crypto';
 import jsYaml from 'js-yaml';
 import { isEqual, uniqWith } from 'lodash';
-import { Op, UniqueConstraintError } from 'sequelize';
+import { col, fn, literal, Op, UniqueConstraintError } from 'sequelize';
 import sharp from 'sharp';
 import * as webpush from 'web-push';
 
@@ -155,7 +155,7 @@ export async function createApp(ctx) {
       result.path = `${path}-${crypto.randomBytes(5).toString('hex')}`;
     }
 
-    const record = await App.create(result);
+    const record = await App.create(result, { raw: true });
 
     ctx.body = getAppFromRecord(record);
     ctx.status = 201;
@@ -166,9 +166,20 @@ export async function createApp(ctx) {
 
 export async function getAppById(ctx) {
   const { appId } = ctx.params;
-  const { App } = ctx.db.models;
+  const { App, AppRating } = ctx.db.models;
 
-  const app = await App.findByPk(appId, { raw: true });
+  const app = await App.findByPk(appId, {
+    raw: true,
+    attributes: {
+      include: [
+        [fn('AVG', col('AppRatings.rating')), 'RatingAverage'],
+        [fn('COUNT', col('AppRatings.AppId')), 'RatingCount'],
+      ],
+      exclude: ['icon', 'style', 'sharedStyle'],
+    },
+    include: [{ model: AppRating, attributes: [] }],
+    group: ['App.id'],
+  });
 
   if (!app) {
     throw Boom.notFound('App not found');
@@ -178,26 +189,47 @@ export async function getAppById(ctx) {
 }
 
 export async function queryApps(ctx) {
-  const { App } = ctx.db.models;
+  const { App, AppRating } = ctx.db.models;
 
   const apps = await App.findAll({
+    attributes: {
+      include: [
+        [fn('AVG', col('AppRatings.rating')), 'RatingAverage'],
+        [fn('COUNT', col('AppRatings.AppId')), 'RatingCount'],
+      ],
+      exclude: ['yaml', 'icon', 'style', 'sharedStyle'],
+    },
     where: { private: false },
+    include: [{ model: AppRating, attributes: [] }],
+    group: ['App.id'],
+    order: [literal('"RatingAverage" DESC NULLS LAST'), ['id', 'ASC']],
     raw: true,
   });
-  ctx.body = apps.map(getAppFromRecord);
+  const ignoredFields = ['yaml'];
+  ctx.body = apps.map(app => getAppFromRecord(app, ignoredFields));
 }
 
 export async function queryMyApps(ctx) {
-  const { App } = ctx.db.models;
+  const { App, AppRating } = ctx.db.models;
   const {
     user: { organizations },
   } = ctx.state;
 
   const apps = await App.findAll({
+    attributes: {
+      include: [
+        [fn('AVG', col('AppRatings.rating')), 'RatingAverage'],
+        [fn('COUNT', col('AppRatings.AppId')), 'RatingCount'],
+      ],
+      exclude: ['yaml', 'icon', 'style', 'sharedStyle'],
+    },
+    include: [{ model: AppRating, attributes: [] }],
+    group: ['App.id'],
+    order: [literal('"RatingAverage" DESC NULLS LAST'), ['id', 'ASC']],
     where: { OrganizationId: { [Op.in]: organizations.map(o => o.id) } },
   });
-
-  ctx.body = apps.map(getAppFromRecord);
+  const ignoredFields = ['yaml'];
+  ctx.body = apps.map(app => getAppFromRecord(app, ignoredFields));
 }
 
 export async function updateApp(ctx) {
@@ -371,6 +403,51 @@ export async function deleteApp(ctx) {
 
   await app.update({ path: null });
   await app.destroy();
+}
+
+export async function getAppRatings(ctx) {
+  const { appId } = ctx.params;
+  const { AppRating, User } = ctx.db.models;
+
+  const ratings = await AppRating.findAll({ where: { AppId: appId }, include: [User], raw: true });
+  ctx.body = ratings.map(({ rating, description, UserId, created, updated, ...r }) => ({
+    rating,
+    description,
+    UserId,
+    name: r['User.name'],
+    $created: created,
+    $updated: updated,
+  }));
+}
+
+export async function submitAppRating(ctx) {
+  const { appId: AppId } = ctx.params;
+  const { App, AppRating, User } = ctx.db.models;
+  const {
+    user: { id: userId },
+  } = ctx.state;
+  const { rating, description } = ctx.request.body;
+
+  const app = await App.findByPk(AppId);
+  const user = await User.findByPk(userId);
+
+  if (!app) {
+    throw Boom.notFound('App not found');
+  }
+
+  const [result] = await AppRating.upsert(
+    { rating, description, UserId: user.id, AppId },
+    { returning: true },
+  );
+
+  ctx.body = {
+    rating,
+    description,
+    UserId: user.id,
+    name: user.name,
+    $created: result.created,
+    $updated: result.updated,
+  };
 }
 
 export async function getAppIcon(ctx) {