+++ packages/studio/actions/db.ts
@@ -1,0 +1,12 @@
+import { IDBPDatabase } from 'idb';
+
+import { AppAction } from './apps';
+
+const initialState: IDBPDatabase = null;
+
+export default (state: IDBPDatabase = initialState, action: AppAction): IDBPDatabase => {
+  switch (action.type) {
+    default:
+      return state;
+  }
+};