+++ packages/studio/components/NavLink/NavLink.tsx
@@ -4,25 +4,26 @@ import { Link, RouteComponentProps } from 'react-router-dom';
 
 interface NavLinkProps {
   className: string;
-  children: React.ReactChild;
-  exact: boolean;
+  children: React.ReactChild[] | React.ReactChild;
+  exact?: boolean;
   to: string;
 }
 
-export default class NavLink extends React.Component<NavLinkProps & RouteComponentProps> {
-  render(): JSX.Element {
-    const { location, className, exact, to, children } = this.props;
-
-    return (
-      <Link
-        className={classNames(className, 'is-radiusless', {
-          'is-active':
-            location.pathname === to || (!exact && location.pathname.startsWith(`${to}/`)),
-        })}
-        to={to}
-      >
-        {children}
-      </Link>
-    );
-  }
+export default function NavLink({
+  location,
+  className,
+  exact,
+  to,
+  children,
+}: NavLinkProps & RouteComponentProps): JSX.Element {
+  return (
+    <Link
+      className={classNames(className, 'is-radiusless', {
+        'is-active': location.pathname === to || (!exact && location.pathname.startsWith(`${to}/`)),
+      })}
+      to={to}
+    >
+      {children}
+    </Link>
+  );
 }